make compute_geometry compute both corners and centroids
